+++ index.js
@@ -5,15 +5,17 @@
 'use strict';
 
 // Module dependencies.
-var client=require('./lib/client');
-var comment=require('./lib/comment');
+const client=require('./lib/client');
+const comment=require('./lib/comment');
 
 // Public interface.
-module.exports={
+const akismet={
   Author: comment.Author,
   Blog: client.Blog,
   Client: client.Client,
   Comment: comment.Comment,
-  CommentType: comment.CommentType,
-  Server: typeof window!='undefined' ? null : require('./lib/server')
+  CommentType: comment.CommentType
 };
+
+module.exports=akismet;
+if(typeof window!='undefined') window.akismet=akismet;