+++ test/client_test.js
@@ -5,133 +5,140 @@
 'use strict';
 
 // Module dependencies.
-var assert=require('assert');
-var env=global.process.env;
-
-var Author=require('../lib/comment').Author;
-var Blog=require('../lib/client').Blog;
-var Client=require('../lib/client').Client;
-var Comment=require('../lib/comment').Comment;
-var CommentType=require('../lib/comment').CommentType;
+const assert=require('assert');
+const clt=require('../lib/client');
+const cmt=require('../lib/comment');
 
 /**
- * Tests the features of the `akismet.Blog` class.
- * @class akismet.tests.BlogTest
- * @static
+ * Tests the features of the `Blog` class.
  */
-var BlogTest={
+class BlogTest {
 
   /**
    * Runs the unit tests.
-   * @method run
    */
-  run: function() {
-    var self=this;
+  run() {
+    let self=this;
     describe('Blog', function() {
+      describe('constructor()', self.testConstructor);
       describe('fromJSON()', self.testFromJSON);
       describe('toJSON()', self.testToJSON);
     });
-  },
+  }
+
+  /**
+   * Tests the constructor.
+   */
+  testConstructor() {
+    it('should initialize the existing properties', () => {
+      let blog=new clt.Blog('https://github.com/cedx/akismet.js', {charset: 'UTF-8', language: 'en'});
+      assert.equal(blog.charset, 'UTF-8');
+      assert.equal(blog.language, 'en');
+      assert.equal(blog.url, 'https://github.com/cedx/akismet.js');
+    });
+
+    it('should not create new properties', () =>
+      assert(!('foo' in new clt.Blog('https://github.com/cedx/akismet.js', {foo: 'bar'})))
+    );
+  }
 
   /**
    * Tests the `fromJSON` method.
-   * @method testFromJSON
    */
-  testFromJSON: function() {
-    it('should return a null reference with a non-object JSON string', function() {
-      assert.strictEqual(Blog.fromJSON('foo'), null);
-    });
+  testFromJSON() {
+    it('should return a null reference with a non-object JSON string', () =>
+      assert.strictEqual(clt.Blog.fromJSON('foo'), null)
+    );
 
-    it('should return an empty instance with an empty JSON object', function() {
-      var blog=Blog.fromJSON('{}');
+    it('should return an empty instance with an empty JSON object', () => {
+      let blog=clt.Blog.fromJSON('{}');
       assert.strictEqual(blog.charset, null);
       assert.strictEqual(blog.language, null);
       assert.strictEqual(blog.url, null);
     });
 
-    it('should return an initialized instance with a non-empty JSON object', function() {
-      var blog=Blog.fromJSON('{ "blog": "http://dev.belin.io/akismet.js", "blog_charset": "UTF-8", "blog_lang": "en" }');
+    it('should return an initialized instance with a non-empty JSON object', () => {
+      let blog=clt.Blog.fromJSON('{ "blog": "https://github.com/cedx/akismet.js", "blog_charset": "UTF-8", "blog_lang": "en" }');
       assert.equal(blog.charset, 'UTF-8');
       assert.equal(blog.language, 'en');
-      assert.equal(blog.url, 'http://dev.belin.io/akismet.js');
+      assert.equal(blog.url, 'https://github.com/cedx/akismet.js');
     });
-  },
+  }
 
   /**
    * Tests the `toJSON` method.
-   * @method testToJSON
    */
-  testToJSON: function() {
-    it('should return an empty JSON object with a newly created instance', function() {
-      assert.equal(new Blog().toJSON(), '{}');
-    });
+  testToJSON() {
+    it('should return an empty JSON object with a newly created instance', () =>
+      assert.equal(new clt.Blog().toJSON(), '{}')
+    );
 
-    it('should return a non-empty JSON object with a initialized instance', function() {
-      var blog=new Blog('http://dev.belin.io/akismet.js', { charset: 'UTF-8', language: 'en' });
-      assert.equal(blog.toJSON(), '{"blog":"http://dev.belin.io/akismet.js","blog_charset":"UTF-8","blog_lang":"en"}');
+    it('should return a non-empty JSON object with a initialized instance', () => {
+      let blog=new clt.Blog('https://github.com/cedx/akismet.js', {charset: 'UTF-8', language: 'en'});
+      assert.equal(blog.toJSON(), '{"blog":"https://github.com/cedx/akismet.js","blog_charset":"UTF-8","blog_lang":"en"}');
     });
   }
-};
+}
 
 /**
- * Tests the features of the `akismet.Client` class.
- * @class akismet.tests.ClientTest
- * @static
+ * Tests the features of the `Client` class.
  */
-var ClientTest={
+class ClientTest {
 
   /**
-   * The client used to query the service database.
-   * @property _client
-   * @type akismet.Client
-   * @private
+   * Initializes a new instance of the class.
    */
-  _client: new Client(
-    env.AKISMET_API_KEY,
-    env.AKISMET_BLOG,
-    { serviceUrl: 'AKISMET_SERVICE_URL' in env ? env.AKISMET_SERVICE_URL : 'https://'+Client.DEFAULT_SERVICE }
-  ),
+  constructor() {
 
-  /**
-   * A comment with content marked as ham.
-   * @property _ham
-   * @type akismet.Comment
-   * @private
-   */
-  _ham: new Comment({
-    author: new Author({
-      ipAddress: '192.168.0.1',
-      name: 'Akismet.js',
-      url: 'http://dev.belin.io/akismet.js',
-      userAgent: 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9) AppleWebKit/537.71 (KHTML, like Gecko) Version/7.0 Safari/537.71'
-    }),
-    content: 'I\'m testing out the Service API.',
-    referrer: 'https://www.npmjs.org/package/akismet-js',
-    type: CommentType.COMMENT
-  }),
+    /**
+     * The client used to query the service database.
+     * @var {Client}
+     * @private
+     */
+    this._client=new clt.Client(
+      process.env.AKISMET_API_KEY,
+      'AKISMET_BLOG' in process.env ? process.env.AKISMET_BLOG : 'https://github.com/cedx/akismet.js',
+      {isTest: true, serviceUrl: 'AKISMET_SERVICE_URL' in process.env ? process.env.AKISMET_SERVICE_URL : 'https://'+clt.Client.DEFAULT_SERVICE}
+    );
 
-  /**
-   * A comment with content marked as spam.
-   * @property _spam
-   * @type akismet.Comment
-   * @private
-   */
-  _spam: new Comment({
-    author: new Author({
-      ipAddress: '127.0.0.1',
-      name: 'viagra-test-123',
-      userAgent: 'Spam Bot/6.6.6'
-    }),
-    content: 'Spam!',
-    type: CommentType.TRACKBACK
-  }),
+    /**
+     * A comment with content marked as ham.
+     * @var {Comment}
+     * @private
+     */
+    this._ham=new cmt.Comment({
+      author: new cmt.Author({
+        ipAddress: '192.168.0.1',
+        name: 'Akismet.js',
+        url: 'https://github.com/cedx/akismet.js',
+        userAgent: 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:42.0) Gecko/20100101 Firefox/42.0'
+      }),
+      content: 'I\'m testing out the Service API.',
+      referrer: 'https://www.npmjs.com/package/akismet-js',
+      type: cmt.CommentType.COMMENT
+    });
+
+    /**
+     * A comment with content marked as spam.
+     * @var {Comment}
+     * @private
+     */
+    this._spam=new cmt.Comment({
+      author: new cmt.Author({
+        ipAddress: '127.0.0.1',
+        name: 'viagra-test-123',
+        userAgent: 'Spam Bot/6.6.6'
+      }),
+      content: 'Spam!',
+      type: cmt.CommentType.TRACKBACK
+    });
+  }
 
   /**
    * Runs the unit tests.
-   * @method run
    */
-  run: function() {
-    var self=this;
+  run() {
+    let self=this;
     describe('Client', function() {
       this.timeout(10000);
       describe('verifyKey()', self.testVerifyKey.bind(self));
@@ -139,78 +146,58 @@ var ClientTest={
       describe('submitSpam()', self.testSubmitSpam.bind(self));
       describe('checkComment()', self.testCheckComment.bind(self));
     });
-  },
+  }
 
   /**
    * Tests the `checkComment` method.
-   * @method testCheckComment
    */
-  testCheckComment: function() {
-    var self=this;
-    it('should return `false` for valid comment (e.g. ham)' , function(done) {
-      self._client.checkComment(self._ham, function(err, res) {
-        assert.ifError(err);
-        assert.strictEqual(res, false);
-        done();
-      });
-    });
+  testCheckComment() {
+    it('should return `false` for valid comment (e.g. ham)' , () =>
+      this._client.checkComment(this._ham).then(res => assert.strictEqual(res, false))
+    );
 
-    it('should return `true` for invalid comment (e.g. spam)' , function(done) {
-      self._client.checkComment(self._spam, function(err, res) {
-        assert.ifError(err);
-        assert.strictEqual(res, true);
-        done();
-      });
-    });
-  },
+    it('should return `true` for invalid comment (e.g. spam)' , () =>
+      this._client.checkComment(this._spam).then(res => assert.strictEqual(res, true))
+    );
+  }
 
   /**
    * Tests the `submitHam` method.
-   * @method testSubmitHam
    */
-  testSubmitHam: function() {
-    var self=this;
-    it('should complete without error' , function(done) {
-      self._client.submitHam(self._ham, done);
-    });
-  },
+  testSubmitHam() {
+    it('should complete without error' , () =>
+      this._client.submitHam(this._ham)
+    );
+  }
 
   /**
    * Tests the `submitSpam` method.
-   * @method testSubmitSpam
    */
-  testSubmitSpam: function() {
-    var self=this;
-    it('should complete without error' , function(done) {
-      self._client.submitSpam(self._spam, done);
-    });
-  },
+  testSubmitSpam() {
+    it('should complete without error' , () =>
+      this._client.submitSpam(this._spam)
+    );
+  }
 
   /**
    * Tests the `verifyKey` method.
-   * @method testVerifyKey
    */
-  testVerifyKey: function() {
-    var self=this;
-    it('should return `true` for a valid API key' , function(done) {
-      self._client.verifyKey(function(err, res) {
-        assert.ifError(err);
-        assert.strictEqual(res, true);
-        done();
-      });
-    });
+  testVerifyKey() {
+    it('should return `true` for a valid API key' , () =>
+      this._client.verifyKey().then(res => assert.strictEqual(res, true))
+    );
 
-    it('should return `false` for an invalid API key' , function(done) {
-      var client=new Client('viagra-test-123', self._client.blog, { serviceUrl: self._client.serviceUrl });
-      client.verifyKey(function(err, res) {
-        assert.ifError(err);
-        assert.strictEqual(res, false);
-        done();
+    it('should return `false` for an invalid API key' , () => {
+      let client=new clt.Client('viagra-test-123', this._client.blog, {
+        isTest: this._client.isTest,
+        serviceUrl: this._client.serviceUrl
       });
+
+      return client.verifyKey().then(res => assert.strictEqual(res, false));
     });
   }
-};
+}
 
 // Run all tests.
-BlogTest.run();
-ClientTest.run();
+new BlogTest().run();
+new ClientTest().run();